requirements.txt
fastapi==0.121.0
numpy==2.3.4
openai==2.7.1
opencv_python==4.10.0.84
opencv_python==4.12.0.88
Pillow==12.0.0
pydantic==2.12.4
python-dotenv==1.2.1
Requests==2.32.5
uvicorn==0.38.0
